Add execution tracing (#9937)
* Add execution tracing
Consolidate indexing to use sd::LongType rather than ints or unsigned.

* Fix javacpp compilation

* Cuda updates

* Remove commented code.

* Clean up more commented code
Migrate  boolean trace ops from validate and execute to execute.

* Fix signature

* Fix signature

* Migrate test to better file
Ensure tests purges trace and toggles off.

* Add small assertions

* Ensure iargs has a null check

* Update ADR
Address nits
Refactor op exec trace check in to 1 method

+++ libnd4j/include/legacy/cpu/NativeOpExecutioner.cpp
@@ -93,7 +93,7 @@ void NativeOpExecutioner::execIndexReduce(sd::LaunchContext *lc, int opNum, cons
                                           const sd::LongType *hXShapeInfo, const void *dX,
                                           const sd::LongType *dXShapeInfo, void *extraParams, void *hZ,
                                           const sd::LongType *hZShapeInfo, void *dZ, const sd::LongType *dZShapeInfo,
-                                          int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
+                                          long long int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
                                           const sd::LongType *tadOffsets) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto zType = sd::ArrayOptions::dataType(hZShapeInfo);
@@ -124,7 +124,7 @@ void NativeOpExecutioner::execBroadcast(sd::LaunchContext *lc, int opNum, const 
                                         const sd::LongType *dXShapeInfo, const void *hY,
                                         const sd::LongType *hYShapeInfo, const void *dY,
                                         const sd::LongType *dYShapeInfo, void *hZ, const sd::LongType *hZShapeInfo,
-                                        void *dZ, const sd::LongType *dZShapeInfo, int *dimension, int dimensionLength,
+                                        void *dZ, const sd::LongType *dZShapeInfo, long long int *dimension, int dimensionLength,
                                         const sd::LongType *tadOnlyShapeInfo, const sd::LongType *tadOffsets,
                                         const sd::LongType *tadOnlyShapeInfoZ, const sd::LongType *tadOffsetsZ) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
@@ -207,7 +207,7 @@ void NativeOpExecutioner::execInverseBroadcast(
     sd::LaunchContext *lc, int opNum, const void *hX, const sd::LongType *hXShapeInfo, const void *dX,
     const sd::LongType *dXShapeInfo, const void *hY, const sd::LongType *hYShapeInfo, const void *dY,
     const sd::LongType *dYShapeInfo, void *hZ, const sd::LongType *hZShapeInfo, void *dZ,
-    const sd::LongType *dZShapeInfo, int *dimension, int dimensionLength, const sd::LongType *tadOnlyShapeInfo,
+    const sd::LongType *dZShapeInfo, long long int *dimension, int dimensionLength, const sd::LongType *tadOnlyShapeInfo,
     const sd::LongType *tadOffsets, const sd::LongType *tadOnlyShapeInfoZ, const sd::LongType *tadOffsetsZ) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto yType = sd::ArrayOptions::dataType(hYShapeInfo);
@@ -249,7 +249,7 @@ void NativeOpExecutioner::execBroadcastBool(sd::LaunchContext *lc, int opNum, co
                                             const sd::LongType *hYShapeInfo, const void *dY,
                                             const sd::LongType *dYShapeInfo, void *hZ, const sd::LongType *hZShapeInfo,
                                             void *dZ, const sd::LongType *dZShapeInfo, void *extraParams,
-                                            int *dimension, int dimensionLength, const sd::LongType *tadOnlyShapeInfo,
+                                            long long int *dimension, int dimensionLength, const sd::LongType *tadOnlyShapeInfo,
                                             const sd::LongType *tadOffsets, const sd::LongType *tadOnlyShapeInfoZ,
                                             const sd::LongType *tadOffsetsZ) {
   if (shape::isEmpty(hXShapeInfo) || shape::isEmpty(hYShapeInfo)) return;
@@ -293,7 +293,7 @@ void NativeOpExecutioner::execInverseBroadcastBool(
     sd::LaunchContext *lc, int opNum, const void *hX, const sd::LongType *hXShapeInfo, const void *dX,
     const sd::LongType *dXShapeInfo, const void *hY, const sd::LongType *hYShapeInfo, const void *dY,
     const sd::LongType *dYShapeInfo, void *hZ, const sd::LongType *hZShapeInfo, void *dZ,
-    const sd::LongType *dZShapeInfo, void *extraParams, int *dimension, int dimensionLength,
+    const sd::LongType *dZShapeInfo, void *extraParams, long long int *dimension, int dimensionLength,
     const sd::LongType *tadOnlyShapeInfo, const sd::LongType *tadOffsets, const sd::LongType *tadOnlyShapeInfoZ,
     const sd::LongType *tadOffsetsZ) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
@@ -327,7 +327,7 @@ void NativeOpExecutioner::execBroadcastInt(
     sd::LaunchContext *lc, int opNum, const void *hX, const sd::LongType *hXShapeInfo, const void *dX,
     const sd::LongType *dXShapeInfo, const void *hY, const sd::LongType *hYShapeInfo, const void *dY,
     const sd::LongType *dYShapeInfo, void *hZ, const sd::LongType *hZShapeInfo, void *dZ,
-    const sd::LongType *dZShapeInfo, int *dimension, int dimensionLength, const sd::LongType *tadOnlyShapeInfo,
+    const sd::LongType *dZShapeInfo, long long int *dimension, int dimensionLength, const sd::LongType *tadOnlyShapeInfo,
     const sd::LongType *tadOffsets, const sd::LongType *tadOnlyShapeInfoZ, const sd::LongType *tadOffsetsZ) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto yType = sd::ArrayOptions::dataType(hYShapeInfo);
@@ -382,7 +382,7 @@ void NativeOpExecutioner::execInverseBroadcastInt(
     sd::LaunchContext *lc, int opNum, const void *hX, const sd::LongType *hXShapeInfo, const void *dX,
     const sd::LongType *dXShapeInfo, const void *hY, const sd::LongType *hYShapeInfo, const void *dY,
     const sd::LongType *dYShapeInfo, void *hZ, const sd::LongType *hZShapeInfo, void *dZ,
-    const sd::LongType *dZShapeInfo, int *dimension, int dimensionLength, const sd::LongType *tadOnlyShapeInfo,
+    const sd::LongType *dZShapeInfo, long long int *dimension, int dimensionLength, const sd::LongType *tadOnlyShapeInfo,
     const sd::LongType *tadOffsets, const sd::LongType *tadOnlyShapeInfoZ, const sd::LongType *tadOffsetsZ) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto yType = sd::ArrayOptions::dataType(hYShapeInfo);
@@ -535,7 +535,7 @@ void NativeOpExecutioner::execReduceFloat(sd::LaunchContext *lc, int opNum, cons
                                           const sd::LongType *hXShapeInfo, const void *dX,
                                           const sd::LongType *dXShapeInfo, void *extraParams, void *hZ,
                                           const sd::LongType *hZShapeInfo, void *dZ, const sd::LongType *dZShapeInfo,
-                                          int *dimension, int dimensionLength) {
+                                          long long int *dimension, int dimensionLength) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto zType = sd::ArrayOptions::dataType(hZShapeInfo);
 
@@ -550,7 +550,7 @@ void NativeOpExecutioner::execReduceSame(sd::LaunchContext *lc, int opNum, const
                                          const sd::LongType *hXShapeInfo, const void *dX,
                                          const sd::LongType *dXShapeInfo, void *extraParams, void *hZ,
                                          const sd::LongType *hZShapeInfo, void *dZ, const sd::LongType *dZShapeInfo,
-                                         int *dimension, int dimensionLength) {
+                                         long long int *dimension, int dimensionLength) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
 
   // nothing to do here if result is empty
@@ -567,7 +567,7 @@ void NativeOpExecutioner::execReduceBool(sd::LaunchContext *lc, int opNum, const
                                          const sd::LongType *hXShapeInfo, const void *dX,
                                          const sd::LongType *dXShapeInfo, void *extraParams, void *hZ,
                                          const sd::LongType *hZShapeInfo, void *dZ, const sd::LongType *dZShapeInfo,
-                                         int *dimension, int dimensionLength) {
+                                         long long int *dimension, int dimensionLength) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto zType = sd::ArrayOptions::dataType(hZShapeInfo);
 
@@ -585,7 +585,7 @@ void NativeOpExecutioner::execReduceLong(sd::LaunchContext *lc, int opNum, const
                                          const sd::LongType *hXShapeInfo, const void *dX,
                                          const sd::LongType *dXShapeInfo, void *extraParams, void *hZ,
                                          const sd::LongType *hZShapeInfo, void *dZ, const sd::LongType *dZShapeInfo,
-                                         int *dimension, int dimensionLength) {
+                                         long long int *dimension, int dimensionLength) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto zType = sd::ArrayOptions::dataType(hZShapeInfo);
 
@@ -719,7 +719,7 @@ void NativeOpExecutioner::execReduce3(sd::LaunchContext *lc, int opNum, const vo
                                       const void *dX, const sd::LongType *dXShapeInfo, void *extraParamsVals,
                                       const void *hY, const sd::LongType *hYShapeInfo, const void *dY,
                                       const sd::LongType *dYShapeInfo, void *hZ, const sd::LongType *hZShapeInfo,
-                                      void *dZ, const sd::LongType *dZShapeInfo, int *dimension, int dimensionLength,
+                                      void *dZ, const sd::LongType *dZShapeInfo, long long int *dimension, int dimensionLength,
                                       const sd::LongType *xTadOnlyShapeInfo, const sd::LongType *xTadOffsets,
                                       const sd::LongType *yTadOnlyShapeInfo, const sd::LongType *yTadOffsets) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
@@ -754,7 +754,7 @@ void NativeOpExecutioner::execReduce3All(sd::LaunchContext *lc, int opNum, const
                                          const sd::LongType *dXShapeInfo, void *extraParamsVals, const void *hY,
                                          const sd::LongType *hYShapeInfo, const void *dY,
                                          const sd::LongType *dYShapeInfo, void *hZ, const sd::LongType *hZShapeInfo,
-                                         void *dZ, const sd::LongType *dZShapeInfo, int *dimension, int dimensionLength,
+                                         void *dZ, const sd::LongType *dZShapeInfo, long long int *dimension, int dimensionLength,
                                          const sd::LongType *xTadShapeInfo, const sd::LongType *xOffsets,
                                          const sd::LongType *yTadShapeInfo, const sd::LongType *yOffsets) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
@@ -780,7 +780,7 @@ void NativeOpExecutioner::execReduce3TAD(sd::LaunchContext *lc, int opNum, const
                                          const sd::LongType *dXShapeInfo, void *extraParamsVals, const void *hY,
                                          const sd::LongType *hYShapeInfo, const void *dY,
                                          const sd::LongType *dYShapeInfo, void *hZ, const sd::LongType *hZShapeInfo,
-                                         void *dZ, const sd::LongType *dZShapeInfo, int *dimension, int dimensionLength,
+                                         void *dZ, const sd::LongType *dZShapeInfo, long long int *dimension, int dimensionLength,
                                          const sd::LongType *tadShapeInfo, const sd::LongType *tadOffsets,
                                          const sd::LongType *yTadShapeInfo, const sd::LongType *yTadOffsets) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
@@ -863,7 +863,7 @@ void NativeOpExecutioner::execScalar(sd::LaunchContext *lc, int opNum, void cons
                                      void const *dX, sd::LongType const *dXShapeInfo, void *extraParams, void *hZ,
                                      sd::LongType const *hZShapeInfo, void *dZ, sd::LongType const *dZShapeInfo,
                                      void const *hScalars, sd::LongType const *hScalarShapeInfo, void const *dScalars,
-                                     sd::LongType const *dScalarShapeInfo, int *dimension, int dimensionLength,
+                                     sd::LongType const *dScalarShapeInfo, long long int *dimension, int dimensionLength,
                                      sd::LongType const *tadShapeInfo, sd::LongType const *tadOffsets,
                                      sd::LongType const *tadShapeInfoZ, sd::LongType const *tadOffsetsZ) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
@@ -935,7 +935,8 @@ void NativeOpExecutioner::execScalarBool(
     sd::LaunchContext *lc, int opNum, const void *hX, const sd::LongType *hXShapeInfo, const void *dX,
     const sd::LongType *dXShapeInfo, void *extraParams, void *hZ, const sd::LongType *hZShapeInfo, void *dZ,
     const sd::LongType *dZShapeInfo, const void *hScalars, const sd::LongType *hScalarShapeInfo, const void *dScalars,
-    const sd::LongType *dScalarShapeInfo, int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
+    const sd::LongType *dScalarShapeInfo,
+                                         long long int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
     const sd::LongType *tadOffsets, const sd::LongType *tadShapeInfoZ, const sd::LongType *tadOffsetsZ) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto yType = sd::ArrayOptions::dataType(hScalarShapeInfo);
@@ -1001,7 +1002,8 @@ void NativeOpExecutioner::execScalarInt(
     sd::LaunchContext *lc, int opNum, const void *hX, const sd::LongType *hXShapeInfo, const void *dX,
     const sd::LongType *dXShapeInfo, void *extraParams, void *hZ, const sd::LongType *hZShapeInfo, void *dZ,
     const sd::LongType *dZShapeInfo, const void *hScalars, const sd::LongType *hScalarShapeInfo, const void *dScalars,
-    const sd::LongType *dScalarShapeInfo, int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
+    const sd::LongType *dScalarShapeInfo,
+                                        long long int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
     const sd::LongType *tadOffsets, const sd::LongType *tadShapeInfoZ, const sd::LongType *tadOffsetsZ) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto yType = sd::ArrayOptions::dataType(hScalarShapeInfo);
@@ -1090,7 +1092,7 @@ void NativeOpExecutioner::execSummaryStats(sd::LaunchContext *lc, int opNum, con
                                            const sd::LongType *hXShapeInfo, const void *dX,
                                            const sd::LongType *dXShapeInfo, void *extraParams, void *hZ,
                                            const sd::LongType *hZShapeInfo, void *dZ, const sd::LongType *dZShapeInfo,
-                                           int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
+                                           long long int *dimension, int dimensionLength, const sd::LongType *tadShapeInfo,
                                            const sd::LongType *tadOffsets, bool biasCorrected) {
   auto xType = sd::ArrayOptions::dataType(hXShapeInfo);
   auto zType = sd::ArrayOptions::dataType(hZShapeInfo);
